Fix local dev environment: service role key + server-side relationships
The dev bypass (DEV_BYPASS_AUTH) only skipped route protection but didn't
create a real Supabase session, so RLS-protected tables returned no data
and client-side writes were silently blocked.

Two fixes:
1. server.ts — when DEV_BYPASS_AUTH=true, use the service role key
   (SUPABASE_SERVICE_ROLE_KEY) instead of the anon key. This bypasses RLS
   for server-side queries while still filtering by user_id manually, so
   only the dev user's own data is returned.

2. relationships page — refactored from a client component (browser Supabase
   queries, blocked by RLS) to a server component + server actions pattern.
   Data fetching and mutations now go through server.ts, which picks up the
   service role key in dev and real auth cookies in prod.
